+++ src/index.tsx
@@ -3,11 +3,18 @@ import MarkdownCMD from './MarkdownCMD';
 import Markdown from './Markdown';
 import type { MarkdownCMDRef, MarkdownRef, IMarkdownMath, ITypedChar, MarkdownProps, MarkdownCMDProps } from './defined';
 
+import Button from './components/Button';
+import CopyButton from './components/CopyButton';
+import DownloadButton from './components/DownloadButton';
+import CodeBlockActions from './components/CodeBlockActions';
+import CodeBlockWrap from './components/CodeBlockWrap';
+
 // 导出 ConfigProvider 相关
 export { ConfigProvider, useConfig, useLocale } from './context/ConfigProvider';
+export { useThemeState } from './context/MarkdownThemeProvider';
 export * from './context/ConfigProvider/mermaid.type';
 export type { Locale, ConfigProviderProps, ConfigContextType } from './context/ConfigProvider';
 
 export default Markdown;
 export type { MarkdownCMDRef, MarkdownRef, IMarkdownMath, ITypedChar, MarkdownProps, MarkdownCMDProps };
-export { Markdown, MarkdownCMD };
+export { Markdown, MarkdownCMD, Button, CopyButton, DownloadButton, CodeBlockActions, CodeBlockWrap };
